\name{globalboosttest-internal}
\alias{perm}


\title{Internal globalboosttest Functions}
\description{
  Internal globalboosttest functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}

